+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- `fill` prop to `PageBlock` and `Box` components.
+
 ## [9.112.28] - 2020-04-02
 
 ### Fixed